Make UnstructuredContent return a copy without mutating contents: - UnstructuredContent now returns a copy of the underlying data and does not modify anything in the underlying object - Any changes to the contents of Unstructured should be made either by directly accessing it's properties or using the provided SetUnstructuredContent method

+++ pkg/printers/internalversion/describe_test.go
@@ -1782,14 +1782,14 @@ URL:	http://localhost
 				"name":              "MyName",
 				"namespace":         "MyNamespace",
 				"creationTimestamp": "2017-04-01T00:00:00Z",
-				"resourceVersion":   123,
+				"resourceVersion":   int64(123),
 				"uid":               "00000000-0000-0000-0000-000000000001",
 				"dummy3":            "present",
 			},
 			"items": []interface{}{
 				map[string]interface{}{
 					"itemBool": true,
-					"itemInt":  42,
+					"itemInt":  int64(42),
 				},
 			},
 			"url":    "http://localhost",
